# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  auth.Permission:
    properties:
      can_export:
        type: boolean
      can_print:
        type: boolean
      function_code:
        type: string
      permission_level:
        type: string
    type: object
  auth.SessionUser:
    properties:
      department_id:
        type: integer
      display_name:
        type: string
      id:
        type: integer
      permissions:
        items:
          $ref: '#/definitions/auth.Permission'
        type: array
      roles:
        items:
          type: string
        type: array
      username:
        type: string
    type: object
...
  auth.User:
    properties:
      departmentID:
        format: int64
        type: integer
      displayName:
        type: string
      id:
        format: int64
        type: integer
      passwordHash:
        type: string
      status:
        type: string
      username:
        type: string
    type: object
...
  handlers.loginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
...
  handlers.swaggerEnvelope:
    type: object
  handlers.swaggerMessageResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
host: localhost:5180
info:
  contact: {}
  description: Backend API for the MI property management system
  title: MI Management Information System API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user and returns a bearer token plus the current
        user profile.
      parameters:
      - description: Login request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.swaggerEnvelope'
            - properties:
                token:
                  type: string
                user:
                  $ref: '#/definitions/auth.User'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
      summary: Log in user
      tags:
      - Auth
  /auth/me:
    get:
      consumes:
      - application/json
      description: Returns the authenticated session user resolved from the bearer
        token.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.swaggerEnvelope'
            - properties:
                user:
                  $ref: '#/definitions/auth.SessionUser'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"